front: move PathStepMarker in the common folder
