+++ front/src/common/Map/components/PathStepMarker.tsx
